Fix memory corruption in deferred lists
This fixes two bugs.

1. memcpy has undefined behavior when copying into overlapping memory
   regions.
2. We accidentally were copying num_deferred_packet bytes instead of
   num_deferred_packet elements

Switch to memmove which has defined behavior for overlapping memory
regions and fix the copy size.

The address sanitizer has been enabled for unit tests. This forces
failures on bad memory access but breaks the use of `ctest -T memcheck`.
For now I think it is more valuable for our tests to force failure on
invalid memory.