Fix build on non-unix systems
clang supports gcc everywhere, gcc supports asan on linux, mingw doesn't
support it. This is likely an overly restrictive condition, but we only
need the asan to run on one set of tests. This is good enough to have
the asan run most of the time but not break the windows build.